Multidisciplinary Project - YOLOv5
This was a Multidisciplinary University Project where as a team we had to build a robot that would navigate through an obstacle course while recognising the images placed on the obstacles. This file contains the code in which YOLOv5 published by Ultralytics, was used as the model to detect the images.